Added webpack, create modules for filters and tasks
