+++ public/index.html
@@ -143,6 +143,6 @@
       </section>
     </main>
 
-    <script src="../src/main.js" type="text/javascript"></script>
+    <script src="./bundle.js" type="text/javascript"></script>
   </body>
 </html>
